==> 01_json_scraping.ipynb <==
import json

file_path = '../../data/json_hangling/spotify_playlist.json'

with open(file_path, 'r') as json_file:
    data = json.load(json_file)


data

data['status']

data['contents']['items'][0]['name'] # write out the first track name

for track in data['contents']['items']: # write out all track name
    print(track['name'])
