==> 05_date_dimension.ipynb <==
import pandas as pd
from datetime import datetime

start_date = datetime(2023, 1, 1)
end_date = datetime(2028, 1, 1)

date_range = pd.date_range(start_date, end_date)

date_df = pd.DataFrame(date_range, columns=["date"])

date_df

date_df["year"] = date_df["date"].dt.year
date_df["month"] = date_df["date"].dt.month
date_df["day"] = date_df["date"].dt.day
date_df["day_of_week"] = date_df["date"].dt.dayofweek + 1
date_df["is_weekend"] = date_df["day_of_week"].isin([6,7])

date_df


date_df.to_csv('date_dimension.csv', index = False)
