==> 07_transform_load.ipynb <==
import requests
import json
import os
import pandas as pd
pd.set_option('display.max_columns', 30)
from datetime import datetime
from dateutil.relativedelta import relativedelta

"""
1. get the data from S3
2. weather data transformations
3. taxi trips transformations - done
4. update payment_type_master - done
5. update company_master - done
6. update taxi_trips with company and payment_type id-s (replace the string values with id-s from the latest master tables) - done
7. upload weather data to S3
8. upload taxi data to S3
9. upload newest payment_type_master and company_master to S3
"""

## taxi_trips transformation codes

current_datetime = datetime.now() - relativedelta(months=2) # 2 months earlier from now
formatted_datetime = current_datetime.strftime("%Y-%m-%d") # formating the 

url = (
    f"https://data.cityofchicago.org/resource/ajtu-isnz.json?"
    f"$where=trip_start_timestamp >= '{formatted_datetime}T00:00:00' "
    f"AND trip_start_timestamp <= '{formatted_datetime}T23:59:59'&$limit=30000"
)

headers = {'X-App-Token' : os.environ.get('CHICAGO_API_TOKEN')}

response = requests.get(url)

data = response.json()

taxi_trips = pd.DataFrame(data)

taxi_trips.drop(["pickup_census_tract", "dropoff_census_tract"], axis=1, inplace=True) # lot of values are NaN values
taxi_trips.drop(["pickup_centroid_location", "dropoff_centroid_location"], axis=1, inplace=True) # duplicated data we don't need them

taxi_trips.dropna(inplace = True)

taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id", "dropoff_community_area": "dropoff_community_area_id"}, inplace=True)

taxi_trips["datetime_for_weather"] = pd.to_datetime(taxi_trips["trip_start_timestamp"]).dt.floor("H")

taxi_trips.head()

### taxi_trips transformation function

def taxi_trips_transformations(taxi_trips: pd.DataFrame) -> pd.DataFrame:
    """
    Transforms a DataFrame containing taxi trip data by performing the following steps:
    
    1. Drops unnecessary columns: "pickup_census_tract", "dropoff_census_tract", 
       "pickup_centroid_location", and "dropoff_centroid_location".
    2. Removes rows with missing values.
    3. Renames columns "pickup_community_area" and "dropoff_community_area" to 
       "pickup_community_area_id" and "dropoff_community_area_id", respectively.
    4. Adds a new column "datetime_for_weather", which is derived by rounding 
       the "trip_start_timestamp" column to the nearest hour.

    Args:
        taxi_trips (pd.DataFrame): The input DataFrame containing taxi trip data.

    Returns:
        pd.DataFrame: The transformed DataFrame with the specified modifications.
    """

    if not isinstance(taxi_trips, pd.DataFrame):
        raise TypeError('taxi_trips is not a valid pandas DataFrame.')

    taxi_trips.drop(["pickup_census_tract", "dropoff_census_tract",
                     "pickup_centroid_location", "dropoff_centroid_location"], axis=1, inplace=True)

    taxi_trips.dropna(inplace = True)

    taxi_trips.rename(columns={"pickup_community_area": "pickup_community_area_id", "dropoff_community_area": "dropoff_community_area_id"}, inplace=True)

    taxi_trips["datetime_for_weather"] = pd.to_datetime(taxi_trips["trip_start_timestamp"]).dt.floor("H")

    return taxi_trips

### company update codes

company_master = taxi_trips['company'].drop_duplicates().reset_index(drop = True)

company_master = pd.DataFrame(
    {
        'company_id' : range(1, len(company_master) + 1),
        'company' : company_master
    }
)

company_master.tail()

new_company_data = [
    {'company' : 'Taxicab Insurance Agency Llc'},
    {'company' : 'X'},
    {'company' : 'Y'}
]

new_company_mapping = pd.DataFrame(new_company_data)

new_company_mapping

company_max_id = company_master['company_id'].max()
company_max_id

# several-line alternative
new_companies_list = []

for company in new_company_mapping['company'].values:
    if company not in company_master['company'].values:
        new_companies_list.append(company)

# one-line alternative
new_companies_list = [company for company in new_company_mapping['company'].values if company not in company_master['company'].values]

new_companies_list

new_companies_df = pd.DataFrame(
    {
    'company_id' : range(company_max_id + 1, company_max_id + len(new_companies_list) + 1),
    'company' : new_companies_list
    }
)

new_companies_df

updated_company_master = pd.concat([company_master, new_companies_df], ignore_index = True)
updated_company_master.tail()

def update_company_master(taxi_trips: pd.DataFrame, company_master: pd.DataFrame) -> pd.DataFrame:
    """
    Updates the company master DataFrame by adding new companies found in the taxi trips DataFrame.

    Args:
        taxi_trips (pd.DataFrame): DataFrame containing taxi trip information, including company names.
        company_master (pd.DataFrame): DataFrame containing the existing company master information 
                                       with columns 'company_id' and 'company'.

    Returns:
        pd.DataFrame: Updated company master DataFrame with new companies added. New companies are assigned
                      unique IDs starting from the maximum company_id in the existing company_master DataFrame.
    """

    company_max_id = company_master['company_id'].max()

    new_companies_list = [company for company in taxi_trips['company'].values if company not in company_master['company'].values]

    new_companies_df = pd.DataFrame({
    'company_id' : range(company_max_id + 1, company_max_id + len(new_companies_list) + 1),
    'company' : new_companies_list
    })

    updated_company_master = pd.concat([company_master, new_companies_df], ignore_index = True)

    return updated_company_master


taxi_trips_company_only = pd.DataFrame({
    'company_id' : [1, 2, 3],
    'company' : ['Taxicab Insurance Agency Llc', 'X', 'Y']
})

taxi_trips_company_only

updated_company_master = update_company_master(taxi_trips = taxi_trips_company_only, company_master = company_master)
updated_company_master.tail()

### payment_type_master codes

payment_type_master = taxi_trips['payment_type'].drop_duplicates().reset_index(drop = True)

payment_type_master = pd.DataFrame(
    {
        'payment_type_id' : range(1, len(payment_type_master) + 1),
        'payment_type' : payment_type_master
    }
)

payment_type_master

taxi_trips_payment_type_only = pd.DataFrame({
    'payment_type_id' : [1, 2, 3],
    'payment_type' : ['Credit Card', 'X', 'Y']
})

taxi_trips_payment_type_only

def update_payment_type_master(taxi_trips: pd.DataFrame, payment_type_master: pd.DataFrame) -> pd.DataFrame:
    """
    Updates the payment type master DataFrame by adding new payment types found in the taxi trips DataFrame.

    Args:
        taxi_trips (pd.DataFrame): DataFrame containing taxi trip information, including payment types.
        payment_type_master (pd.DataFrame): DataFrame containing the existing payment type master information 
                                            with columns 'payment_type_id' and 'payment_type'.

    Returns:
        pd.DataFrame: Updated payment type master DataFrame with new payment types added. 
                      New payment types are assigned unique IDs starting from the maximum payment_type_id 
                      in the existing payment_type_master DataFrame.
    """

    payment_type_max_id = payment_type_master['payment_type_id'].max()

    new_payment_types_list = [payment_type for payment_type in taxi_trips['payment_type'].values if payment_type not in payment_type_master['payment_type'].values]

    new_payment_types_df = pd.DataFrame({
    'payment_type_id' : range(payment_type_max_id + 1, payment_type_max_id + len(new_payment_types_list) + 1),
    'payment_type' : new_payment_types_list
    })

    updated_payment_type_master = pd.concat([payment_type_master, new_payment_types_df], ignore_index = True)

    return updated_payment_type_master

updated_payment_type_master = update_payment_type_master(taxi_trips = taxi_trips_payment_type_only, payment_type_master = payment_type_master)
updated_payment_type_master

### creating a generic update master table function

def update_master(taxi_trips: pd.DataFrame, master: pd.DataFrame, id_coulomn: str, value_column: str) -> pd.DataFrame:
    """
    Updates a master DataFrame by adding new values found in the taxi trips DataFrame.

    Args:
        taxi_trips (pd.DataFrame): DataFrame containing taxi trip information, including the column with new values.
        master (pd.DataFrame): DataFrame representing the master data to be updated, including specified ID and value columns.
        id_coulomn (str): Name of the column in the master DataFrame representing unique identifiers.
        value_column (str): Name of the column in the master DataFrame containing the values to be checked and updated.

    Returns:
        pd.DataFrame: Updated master DataFrame with new values added. New values are assigned unique IDs starting 
                      from the maximum value in the specified ID column of the existing master DataFrame.
    """

    max_id = master[id_coulomn].max()

    new_values_list = [value for value in taxi_trips[value_column].values if value not in master[value_column].values]

    new_values_df = pd.DataFrame({
    id_coulomn : range(max_id + 1, max_id + len(new_values_list) + 1),
    value_column : new_values_list
    })

    updated_master = pd.concat([master, new_values_df], ignore_index = True)

    return updated_master

test_payment_type_master = update_master(taxi_trips = taxi_trips_payment_type_only, master = payment_type_master, id_coulomn = 'payment_type_id', value_column = 'payment_type')
test_payment_type_master

test_company_master = update_master(taxi_trips = taxi_trips_company_only, master = company_master, id_coulomn = 'company_id', value_column = 'company')
test_company_master

### update taxi_trips with the most recent company_master and payment_type_master function

def update_taxi_trips_with_master_data(taxi_trips: pd.DataFrame, payment_type_master: pd.DataFrame, company_master: pd.DataFrame) -> pd.DataFrame:
    """
    Updates the taxi trips DataFrame by merging it with payment type and company master DataFrames to replace 
    textual payment type and company data with their corresponding IDs.

    Args:
        taxi_trips (pd.DataFrame): DataFrame containing taxi trip information, including 'payment_type' and 'company' columns.
        payment_type_master (pd.DataFrame): DataFrame containing master data for payment types, with columns 
                                            'payment_type_id' and 'payment_type'.
        company_master (pd.DataFrame): DataFrame containing master data for companies, with columns 
                                       'company_id' and 'company'.

    Returns:
        pd.DataFrame: Updated taxi trips DataFrame where 'payment_type' and 'company' columns are replaced with 
                      'payment_type_id' and 'company_id' columns from the respective master DataFrames.
    """
    taxi_trips_id = taxi_trips.merge(payment_type_master, on = 'payment_type')

    taxi_trips_id = taxi_trips_id.merge(company_master, on = 'company')

    taxi_trips_id.drop(['payment_type', 'company'], axis = 1, inplace = True)

    return taxi_trips_id

taxi_trips_id = update_taxi_trips_with_master_data(taxi_trips = taxi_trips, payment_type_master = payment_type_master, company_master = company_master)

taxi_trips_id.sample(5)

### weather transformations function

def transform_weather_data(weather_data: json) -> pd.DataFrame:
    """
    Transforms weather data in JSON format into a structured pandas DataFrame for easier analysis.

    Args:
        weather_data (json): JSON object containing hourly weather data, including time, temperature, wind speed, 
                             rain, and precipitation.

    Returns:
        pd.DataFrame: A DataFrame with the following columns:
                      - 'datetime': Timestamps converted to datetime objects.
                      - 'tempretaure': Hourly temperatures (in °C or relevant unit).
                      - 'wind_speed': Hourly wind speed (in m/s or relevant unit).
                      - 'rain': Hourly rainfall data.
                      - 'precipitation': Hourly precipitation data.
    """
    weather_data_filtered = {
        "datetime": weather_data["hourly"]["time"],
        "tempretaure": weather_data["hourly"]["temperature_2m"],
        "wind_speed": weather_data["hourly"]["wind_speed_10m"],
        "rain": weather_data["hourly"]["rain"],
        "precipitation": weather_data["hourly"]["precipitation"],
    }

    weather_df = pd.DataFrame(weather_data_filtered)

    weather_df["datetime"] = pd.to_datetime(weather_df["datetime"])

    return weather_df

current_datetime = datetime.now() - relativedelta(months=2)
formatted_datetime = current_datetime.strftime("%Y-%m-%d")

url = "https://archive-api.open-meteo.com/v1/era5"

params = {
    "latitude": 41.85,
    "longitude": -87.65,
    "start_date": formatted_datetime,
    "end_date": formatted_datetime,
    "hourly": "temperature_2m,wind_speed_10m,rain,precipitation"
}

response = requests.get(url, params = params)
weather_data = response.json()

weather_data_df = transform_weather_data(weather_data = weather_data)

weather_data_df
