==> 08_local_visualisations.ipynb <==
from io import StringIO
import boto3
import pandas as pd
import os

pd.set_option('display.max_columns', 50)

aws_access_key_id = os.getenv('AWS_access_key')
aws_secret_access_key = os.getenv('AWS_secret_access_key')

def read_csv_from_s3(bucket: str, path: str, filename: str) -> pd.DataFrame:
    """
    Reads a CSV file from an S3 bucket and returns its contents as a pandas DataFrame.

    Args:
        bucket (str): The name of the S3 bucket where the file is stored.
        path (str): The path or prefix within the bucket where the file is located.
        filename (str): The name of the file to be read.

    Returns:
        pd.DataFrame: A pandas DataFrame containing the content of the CSV file.

    Raises:
        botocore.exceptions.ClientError: If there is an error accessing the S3 bucket or file.
        UnicodeDecodeError: If the file cannot be decoded as UTF-8.
        pd.errors.EmptyDataError: If the file is empty or cannot be parsed as CSV.
    """
    s3 = boto3.client('s3', aws_access_key_id = aws_access_key_id, aws_secret_access_key = aws_secret_access_key)
    
    full_path = f'{path}{filename}'

    object = s3.get_object(Bucket = bucket, Key = full_path)
    object = object['Body'].read().decode('utf-8')
    output_df = pd.read_csv(StringIO(object))

    return output_df

s3 = boto3.client('s3', aws_access_key_id = aws_access_key_id, aws_secret_access_key = aws_secret_access_key)
bucket = 'cubix-chicago-taxi-ld'

community_areas_path = 'transformed_data/community_areas/'
company_path = 'transformed_data/company/'
date_path = 'transformed_data/date/'
payment_type_path = 'transformed_data/payment_type/'
taxi_trips_path = 'transformed_data/taxi_trips/'
weather_path = 'transformed_data/weather/'

community_areas = read_csv_from_s3(bucket = bucket, path = community_areas_path, filename = 'community_areas_master.csv')
company = read_csv_from_s3(bucket = bucket, path = company_path, filename = 'company_master.csv')
date = read_csv_from_s3(bucket = bucket, path = date_path, filename = 'date_dimension.csv')
payment_type = read_csv_from_s3(bucket = bucket, path = payment_type_path, filename = 'payment_type_master.csv')

trips_list = []
weather_list = []

for file in s3.list_objects(Bucket = bucket, Prefix = taxi_trips_path)['Contents']:
    taxi_trip_key = file['Key']

    if taxi_trip_key.split('/')[-1].strip() != '':
        if taxi_trip_key.split('.')[1] == 'csv':

            filename = taxi_trip_key.split('/')[-1]
            trip = read_csv_from_s3(bucket = bucket, path = taxi_trips_path, filename = filename)

            trips_list.append(trip)
            print(f'{filename} has been added.')


trips = pd.concat(trips_list, ignore_index = True)

trips.shape

for file in s3.list_objects(Bucket = bucket, Prefix = weather_path)['Contents']:
    weather_key = file['Key']

    if weather_key.split('/')[-1].strip() != '':
        if weather_key.split('.')[1] == 'csv':

            filename = weather_key.split('/')[-1]
            weather_daily = read_csv_from_s3(bucket = bucket, path = weather_path, filename = filename)

            weather_list.append(weather_daily)
            print(f'{filename} has been added.')


weather = pd.concat(weather_list, ignore_index = True)

# join all the dataframes together

trips_full = pd.merge(trips, weather, left_on = 'datetime_for_weather', right_on = 'datetime', how = 'inner')
trips_full = trips_full.drop(columns = ['datetime'])

trips_full = pd.merge(trips_full, company, left_on = 'company_id', right_on = 'company_id', how = 'inner')
trips_full = trips_full.drop(columns = ['company_id'])

trips_full = pd.merge(trips_full, payment_type, left_on = 'payment_type_id', right_on = 'payment_type_id', how = 'inner')
trips_full = trips_full.drop(columns = ['payment_type_id'])

trips_full = pd.merge(trips_full, community_areas, left_on = 'pickup_community_area_id', right_on = 'area_code', how = 'inner')
trips_full = trips_full.drop(columns = ['pickup_community_area_id', 'area_code'])
trips_full.rename(columns = {'community_name':'pickup_community_area_name'}, inplace = True)

trips_full = pd.merge(trips_full, community_areas, left_on = 'dropoff_community_area_id', right_on = 'area_code', how = 'inner')
trips_full = trips_full.drop(columns = ['dropoff_community_area_id', 'area_code'])
trips_full.rename(columns = {'community_name':'dropoff_community_area_name'}, inplace = True)

date['date'] = pd.to_datetime(date['date'])
trips_full['trip_start_timestamp'] = pd.to_datetime(trips_full['trip_start_timestamp'])

trips_full['trip_start_date'] = trips_full['trip_start_timestamp'].dt.date
trips_full['trip_start_date'] = pd.to_datetime(trips_full['trip_start_date'])

trips_full = pd.merge(trips_full, date, left_on = 'trip_start_date', right_on = 'date', how = 'inner')
trips_full = trips_full.drop(columns = ['date'])

trips_full.head(2)
